books_api/app/data/db/postgres.py
import psycopg2
from psycopg2.extras import RealDictCursor
from app.core import config

def get_connection():
    conn = psycopg2.connect(
        host=config.DB_HOST,
        port=config.DB_PORT,
        dbname=config.DB_NAME,
        user=config.DB_USER,
        password=config.DB_PASSWORD,
        cursor_factory=RealDictCursor
    )
    return conn
