books_api/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0,<0.20.0
pydantic>=1.8.0,<2.0.0
python-multipart>=0.0.5,<0.0.6
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0  # Pour la compatibilité avec PostgreSQL (au cas où)
pytest>=6.2.5,<7.0.0
httpx>=0.19.0,<0.20.0
